Internal function: select basis functions

Internal function: select basis functions

selectBasis(
  data,
  loc,
  D = NULL,
  maxit = 50,
  avgtol = 1e-06,
  max_rank = NULL,
  sequence_rank = NULL,
  method = c("fast", "EM"),
  num_neighbors = 3,
  max_knot = 5000,
  DfromLK = NULL,
  Fk = NULL
)

Arguments

data

An n by T data matrix (NA allowed) with \(z[t]\) as the t-th column.

loc

n by d matrix of coordinates corresponding to n locations.

D

A diagonal matrix.

maxit

An integer for the maximum number of iterations used in indeMLE.

avgtol

A numeric for average tolerance used in indeMLE.

max_rank

An integer of the maximum of K values.

sequence_rank

An array of K values

method

A character of a list of characters.

num_neighbors

An integer.

max_knot

An integer for the maximum number of knots

DfromLK

An n by n diagonal matrix.

Fk

An n by K matrix of basis function values with each column being a basis function taken values at loc.

Value

An mrts object with 6 attributes